requirements.txt
asgiref
autopep8
django
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
pillow
sweetify
djangorestframework
markdown
django-filter
django-crispy-forms
dbbackup
django-ckeditor
pyfcm
html2text
PyPDF2
qr_code
pyshorteners
cryptography